+++ include/gl.h
@@ -28,8 +28,10 @@ public:
 
     Shader();
     ~Shader();
+
     void add(size_t, const char*) const;
     void activate() const;
+
     GLint uniform(const char*) const;
     void bind_attrib(const vector<const char*>&) const;
 };
@@ -48,6 +50,7 @@ public:
 
     Window(size_t, size_t, size_t, size_t, const char*);
     ~Window();
+
     bool alive() const;
-    void render(const Voxel&, const Fluid&, const GLint, const GLint, const double) const;
+    void render(const Voxel&, const Fluid&, const GLint, const GLint, const float) const;
 };
