+++ include/main.h
@@ -43,11 +43,11 @@ const     char*  name   = "smoke";
 constexpr size_t dim_x = 100;
 constexpr size_t dim_y = 100;
 constexpr size_t dim_z = 5;
-constexpr double dx    = 1.0 / 100;
+constexpr float dx     = 1.0 / 100;
 
 using clk                  = std::chrono::system_clock;
 constexpr uint64_t fps     = 3;
 constexpr uint64_t t_unit  = clk::period::den / clk::period::num;
 constexpr auto     t_frame = clk::duration(static_cast<uint64_t>((1.0 / fps) * t_unit));
 
-constexpr double   max_quantity = 1000;
+constexpr float max_quantity = 1000;
